+++ cktap/compat.py
@@ -24,7 +24,6 @@ def sha256s(msg):
     from hashlib import sha256
     return sha256(msg).digest()
 
-@staticmethod
 def hash160(x):
     # classic bitcoin nested hashes
     from .ripemd import RIPEMD160
@@ -65,6 +64,7 @@ def CT_bip32_derive(chain_code, master_priv_pub, subkey_path):
 
 try:
     # Wally Core <https://wally.readthedocs.io/en/release_0.8.3/crypto/>
+    raise ImportError
     import wallycore
 
     from .wrap_wally import hash160, sha256s